wardrobe-backend/ai-service/pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build_backend = "setuptools.build_meta"

[project]
name = "digital-wardrobe-ai-service"
version = "0.1.0"
description = "AI service for generating outfit recommendations"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "google-genai>=0.1.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    "requests>=2.31.0",
    "urllib3>=1.26.0,<2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "ruff>=0.3.0",
    "mypy>=1.9.0",
]

[tool.ruff]
line-length = 88
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "UP", "B"]
ignore = ["E501"]

[tool.mypy]
python_version = "3.10"
strict = true
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
